requirements.txt
hdfs
pyspark==3.1.2
delta-spark